Diagrams become inconsistent if they are applied on an updated model #9
Remove option to display diagram from Class Group (Add a statement to explain where to do it).
Add a menu point in the context menu of the canvas that gives the name of the current Moose ID.
Prevent errors when the MooseModel changed. In that case it is needed to read the data of a Layout again from file. The user is forced to do this explicitely.